refactor

+++ models/user_online_status.go
@@ -5,7 +5,7 @@ import (
 )
 
 type UserOnlineStatus struct {
-	UserId     int       `gorm:"primary_key" json:"user_id"`
+	UserId     uint64    `gorm:"primary_key" json:"user_id"`
 	IsOnline   int       `json:"is_online"`
 	LastOnline time.Time `json:"last_online"`
 }
